requirements.txt
fastmcp
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
watchdog
GitPython
